+++ NEWS.md
@@ -10,6 +10,10 @@
 
 * `ddbs_boundary()`: returns the boundary of geometries.
 
+* `ddbs_concave_hull()`: new function to create the concave hull enclosing a geometry.
+
+* `ddbs_convex_hull()`: new function to create the convex hull enclosing a geometry.
+
 * `ddbs_create_conn()`: new convenient function to create a DuckDB connection.
 
 * `ddbs_drivers()`: get list of GDAL drivers and file formats supported by DuckDB spatial extension.